+++ opentelemetry-sdk/CHANGELOG.md
@@ -17,6 +17,7 @@
         .with_resource(Resource::empty())
         .build();
     ```
+  - `logs::LogData` struct is deprecated, and scheduled to be removed from public API in `v0.28.0`.
 
 ## 0.27.0